Switch samples to Spring Boot 3

+++ samples/webflux-websocket/src/main/java/com/example/webflux/WebfluxWebsocketApplication.java
@@ -65,7 +65,7 @@ public class WebfluxWebsocketApplication {
 			DatabaseClient dbc,
 			CustomerRepository customerRepository) {
 		return args -> {
-			Mono<Integer> ddl = dbc
+			Mono<Long> ddl = dbc
 					.sql("create table customer(id serial primary key,name varchar (255) not null)")
 					.fetch()
 					.rowsUpdated();
